requirements.txt
# Core Requirements
Flask==2.3.2
gunicorn==20.1.0

# URL & Network Analysis
requests==2.31.0
urllib3==2.0.4
beautifulsoup4==4.12.2
python-whois==0.8.0
tldextract==3.4.4
socketIO-client==0.7.2

# Security & SSL
pyOpenSSL==23.2.0
cryptography==41.0.3
sslscan==1.8.2

# Data Processing & Utilities
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2

# Machine Learning (optional for future enhancements)
scikit-learn==1.3.0
joblib==1.3.2
nltk==3.8.1

# Development & Testing
pytest==7.4.0
Faker==18.11.2
coverage==7.3.0

# Production
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
